Can now connect to XMPP over Websockets instead of just BOSH
This was accomplished by switching to the Gordin branch of Strophe.js
(see https://github.com/Gordin/strophejs, websockets branch)

Note that the config.json format has changed. The "bosh" key is no longer
used. Instead the "xmpp" key now has a "url" subkey. This should specify
either a BOSH url or a Websockets URL.

For example, to use Websockets:

  "xmpp": {
    "domain": "badger.encorelab.org",
    "port": 5222,
    "url": "ws://badger.encorelab.org:5281/"
  },

And to use BOSH:

  "xmpp": {
    "domain": "badger.encorelab.org",
    "port": 5222,
    "url": "http://badger.encorelab.org:5280/"
  },

For BOSH connections, reverse proxying is used (via the local server.js or
Apache/nginx in production), so the URL used localy is always just "/http-bind".
The URL specified in the config is used by server.js to make the reverse proxy
mapping. In an Apache/nginx deployment, this configuration option is not used at all
other than to check whether BOSH or Websockets should be used (if the URL starts with ws,
Sail will use Websockets, otherwise it will use BOSH)

For Websockets, the connections are done directly cross-domain by the browser to the specified URL.

TODO: automatically fall back to BOSH when websockets are not available